Bug fixes for the C++Builder Athens Patch
Mixing try...catch and try...finally is not allowed and will no longer compile after the patch. In these cases, try...finally is now replaced by a scope guard, OnLeavingScope. Additionally, the use of DELPHICLASS has been removed, and the type Longword, which makes the code independent of the platform's bit architecture (32-bit or 64-bit), is now used instead.